requirements.txt
streamlit
ultralytics
torch
torchvision
numpy
pillow
requests
supabase